made changes in folder structure, moved factory to top level
