+++ battery/battery.py
@@ -3,5 +3,5 @@ from abc import ABC, abstractmethod
 
 class Battery(ABC):
     @abstractmethod
-    def battery_should_be_serviced(self):
+    def needs_service(self):
         pass
