+++ battery/nubbin_battery.py
@@ -7,7 +7,7 @@ class NubbinBattery(Battery):
         self.last_service_date = last_service_date
 
     # should be service every 4 years
-    def battery_should_be_serviced(self) -> bool:
+    def needs_service(self) -> bool:
         today = datetime.today().date()
         service_threshold_date = self.last_service_date.replace(year=self.last_service_date.year + 4)
         return service_threshold_date < today
