+++ battery/spindler_battery.py
@@ -7,7 +7,7 @@ class SpindlerBattery(Battery):
         self.last_service_date = last_service_date
 
     # should be service every two years
-    def battery_should_be_serviced(self) -> bool:
+    def needs_service(self) -> bool:
         service_threshold_date = self.last_service_date.replace(year=self.last_service_date.year + 2)
 
         return service_threshold_date < datetime.today().date()
