+++ engine/willoughby_engine.py
@@ -1,4 +1,3 @@
-
 from engine.engine import Engine
 
 
@@ -9,5 +8,5 @@ class WilloughbyEngine(Engine):
         self.__last_service_mileage = last_service_mileage
 
     # serviced every 60k miles
-    def engine_should_be_serviced(self) -> bool:
+    def needs_service(self) -> bool:
         return self.__current_mileage - self.__last_service_mileage > 60000
