+++ test/test_car.py
@@ -2,7 +2,7 @@ import unittest
 from datetime import datetime
 
 
-from car_factory.car_factory import CarFactory
+from car_factory import CarFactory
 
 
 class TestCalliope(unittest.TestCase):
